backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
networkx
pillow
jinja2
httpx

# Google AI — Gemini Vision (floor plan parsing)
google-genai

# Live Navigation modules
# NOTE: Large packages (~1GB combined). Set DETECTION_MODE=mock on free Render tier.
# Set DETECTION_MODE=real on a paid instance with 2GB+ RAM.
ultralytics
easyocr
